+++ carp_context_package/lib/location_datum.dart
@@ -10,8 +10,7 @@ part of context;
 /// Holds location information using the GPS format.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class LocationDatum extends CARPDatum {
-  static const DataFormat CARP_DATA_FORMAT =
-      DataFormat(NameSpace.CARP, DataType.LOCATION);
+  static const DataFormat CARP_DATA_FORMAT = DataFormat(NameSpace.CARP, ContextSamplingPackage.LOCATION);
   DataFormat get format => CARP_DATA_FORMAT;
 
   LocationDatum() : super();
@@ -25,8 +24,7 @@ class LocationDatum extends CARPDatum {
         speedAccuracy = map['speedAccuracy'],
         super();
 
-  factory LocationDatum.fromJson(Map<String, dynamic> json) =>
-      _$LocationDatumFromJson(json);
+  factory LocationDatum.fromJson(Map<String, dynamic> json) => _$LocationDatumFromJson(json);
   Map<String, dynamic> toJson() => _$LocationDatumToJson(this);
 
   /// Latitude in GPS coordinates.