+++ carp_audio_package/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.2
+* Update to `carp_mobile_sensing` v. 0.3.7
+* Using the data types in the package (`audio`, `noise`)
+* Update of readme file.
+
 ## 0.3.1
 * Documentation