+++ carp_communication_package/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.3.2
+* Documentation
+
 ## 0.3.1
 * Documentation